LIME-233: * Remove modal window
